requirements.txt
# python 3.8 (p3.9 not supported because of an unmaintained dependency)
# create the env:
# conda create -n <name> ipython python=3.8 -y
#
# conda install pytorch cudatoolkit=10.1 -c pytorch

torch>=1.5
torchaudio>=0.7.0
hydra-core>=1.1
julius
mne>=0.24.1
pandas>=1.1
python-Levenshtein
scikit-learn
wordfreq
pytest
pytest-cov>=2.6.1
flake8
mypy
numpy
types-PyYAML
spacy<2.4.0
wandb
numba
mne_bids  # for gwilliams2022 dataset
osfclient
matplotlib >= 3.5.1
scipy
submitit
dora-search
flashy
transformers >= 4.15.0 # for BERT and Wav2Vec embeddings
pyyaml
autoreject
